Fixed Issue #60: isnan non-floating-point argument
When compiling with recent versions of gcc (libc6 2.23), the compiler is complaining about a non-floating point argument for builtin isnan function.
Here, the traceback was:
src/scaler.cpp:297:7: error: non-floating-point argument in call to function '__builtin_isnan'

+++ src/scaler.cpp
@@ -294,7 +294,7 @@ void _scale_rgb(DEST& dest,
 		scale.set_bg( it() );
 	    } else {
 		val = interpolate(src, tr, p);
-		if (isnan(val)) {
+		if (isnan((float) val)) {
 		    scale.set_bg( it() );
 		} else {
 		    it() = scale.eval(val);
